+++ packages/better-translation/CHANGELOG.md
@@ -1,5 +1,23 @@
 # better-translation
 
+## 0.6.8
+
+### Patch Changes
+
+- [`4313595`](https://github.com/NoQuarterTeam/better-translation/commit/4313595046161dba304dbd29bc89d7a6e73f618a) Thanks [@JClackett](https://github.com/JClackett)! - Emit the hot Locale values runtime module used by the packaged Svelte provider.
+
+## 0.6.7
+
+### Patch Changes
+
+- [`3261373`](https://github.com/NoQuarterTeam/better-translation/commit/3261373debb159b6a25feab01050a9830e67cd78) Thanks [@JClackett](https://github.com/JClackett)! - Link the package README to the end-to-end Vite plugin demo.
+
+## 0.6.6
+
+### Patch Changes
+
+- [`c92d1d2`](https://github.com/NoQuarterTeam/better-translation/commit/c92d1d27062a47f6d00863f6ee53a6032a823726) Thanks [@JClackett](https://github.com/JClackett)! - Improve npm discoverability with searchable package metadata, a focused README, and bundled license information.
+
 ## 0.6.5
 
 ### Patch Changes